Move Rule to Entry/ folder
