+++ src/Plugin.php
@@ -2,6 +2,7 @@
 namespace Trendwerk\AcfForms;
 
 use Trendwerk\AcfForms\Entry\Entries;
+use Trendwerk\AcfForms\Entry\Rule;
 use Trendwerk\AcfForms\Form\Forms;
 
 final class Plugin
